requirements.txt
streamlit
python-dotenv
langchain_core
langchain_community
langchain_openai
faiss-cpu
pypdf
bs4
chromadb
